Refactored NweetFactory.
